backend/requirements.txt
numpy
praw
pandas
prophet
fastapi
uvicorn         
python-dotenv
pydantic
